+++ otel/tests/auto/autoinstrument-internal.phpt
@@ -6,10 +6,10 @@ otel
 OTEL_TRACES_EXPORTER=console
 --FILE--
 <?php
-var_dump(str_contains('hello', 'el'));
+var_dump(phpversion());
 ?>
 --EXPECTF--
-bool(true)
+string(5) "%s"
 Spans
 Resource
 %A
@@ -17,7 +17,7 @@ Span #0
 	Instrumentation Scope
 		Name         : "php-auto-instrumentation"
 
-	Name        : <global>::str_contains
+	Name        : <global>::phpversion
 	TraceId     : %s
 	SpanId      : %s
 	TraceFlags  : TraceFlags(1)
